// data/github/current/ui/packages/repository-milestone/__tests__/__generated__/MilestoneListTestQuery.graphql.ts
export type MilestoneListTestQuery$variables = Record<PropertyKey, never>;
export type MilestoneListTestQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"MilestoneList">;
  } | null | undefined;
};
export type MilestoneListTestQuery = {
  response: MilestoneListTestQuery$data;
  variables: MilestoneListTestQuery$variables;
};